+++ src/main/java/com/ciaranmckenna/readinglistapp/dto/AuthorRecord.java
@@ -1,3 +1,3 @@
 package com.ciaranmckenna.readinglistapp.dto;
 
-public record AuthorRecord(int id, String firstName, String lastName) {}
+public record AuthorRecord(Long id, String firstName, String lastName) {}
